Add or remove the 'pgmonitor' role for the pgnodemx extension
Per the pgnodemx documentation:
"PostgreSQL version 9.5 or newer is required.
On PostgreSQL version 9.6 or earlier, a role called
pgmonitor must be created, and the user calling these
functions must be granted that role."

As such, the 'pgmonitor' role has been added to the
'setup_pg95.sql' and 'setup_pg96.sql' exporter scripts.
Please note, no such change was added to the PG 9.4
equivalent due to incompatibility with the pgnodemx extension.

If that role is present in PostgreSQL version 10, 11 or 12,
due to an upgrade, it will be removed in favor of the
pg_monitor role.