Updates for postgres-operator integration (#202)
* Data Disk Query Update

An issue was discovered where monitor.proc_diskstats() is not
returning /pgdata when using NFS backed PVs. To correct this,
this update takes the original ccp_nodemx_data_disk query
used by the exporter and places it under a new heading,
ccp_nodemx_disk_activity. A new version of ccp_nodemx_data_disk
is then created that does not join monitor.proc_mountinfo() with
monitor.proc_diskstats(), allowing the /pgdata mount point to
be returned.

The corresponding change to pod_details.json points the 'Disk
Activity' panel to the new 'ccp_nodemx_disk_activity' metric name.

* Container Alert Rules Update

This change adds a custom pg alerting rules files that can be used in
containers. The only difference between this file and the
crunchy-alert-rules-pg.yml.example is that disk size alerts are updated to use
nodemx queries and alert if 75% or 90% of the disk is used.

* Update prometheus/alert-rules.d/crunchy-alert-rules-pg.yml.containers.example

Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>

* Update prometheus/alert-rules.d/crunchy-alert-rules-pg.yml.containers.example

Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>

Co-authored-by: TJ Moore <tj.moore@crunchydata.com>
Co-authored-by: Keith Fiske <keith@keithf4.com>
Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>