Development to Main for v4.4-RC5 (#204)
* Postgres exporter configuration updates related to the pgnodemx extension (#199)

* Add or remove the 'pgmonitor' role for the pgnodemx extension

Per the pgnodemx documentation:
"PostgreSQL version 9.5 or newer is required.
On PostgreSQL version 9.6 or earlier, a role called
pgmonitor must be created, and the user calling these
functions must be granted that role."

As such, the 'pgmonitor' role has been added to the
'setup_pg95.sql' and 'setup_pg96.sql' exporter scripts.
Please note, no such change was added to the PG 9.4
equivalent due to incompatibility with the pgnodemx extension.

If that role is present in PostgreSQL version 10, 11 or 12,
due to an upgrade, it will be removed in favor of the
pg_monitor role.

* queries_nodemx.yml usage fs_type usage update

The fs_type usage value for ccp_nodemx_data_disk causes an
error in the relevant logs:

"Could not parse string: strconv.ParseFloat: parsing \"xfs\": invalid syntax" source="postgres_exporter.go:739"
"Unexpected error parsing column:  ccp_nodemx_data_disk fs_type xfs\n" source="postgres_exporter.go:1361"

To correct this, the fs_type usage value has been updated to
"LABEL" in place of "GAUGE".

* Updates for postgres-operator integration (#202)

* Data Disk Query Update

An issue was discovered where monitor.proc_diskstats() is not
returning /pgdata when using NFS backed PVs. To correct this,
this update takes the original ccp_nodemx_data_disk query
used by the exporter and places it under a new heading,
ccp_nodemx_disk_activity. A new version of ccp_nodemx_data_disk
is then created that does not join monitor.proc_mountinfo() with
monitor.proc_diskstats(), allowing the /pgdata mount point to
be returned.

The corresponding change to pod_details.json points the 'Disk
Activity' panel to the new 'ccp_nodemx_disk_activity' metric name.

* Container Alert Rules Update

This change adds a custom pg alerting rules files that can be used in
containers. The only difference between this file and the
crunchy-alert-rules-pg.yml.example is that disk size alerts are updated to use
nodemx queries and alert if 75% or 90% of the disk is used.

* Update prometheus/alert-rules.d/crunchy-alert-rules-pg.yml.containers.example

Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>

* Update prometheus/alert-rules.d/crunchy-alert-rules-pg.yml.containers.example

Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>

Co-authored-by: TJ Moore <tj.moore@crunchydata.com>
Co-authored-by: Keith Fiske <keith@keithf4.com>
Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>

* Fix pgBackRest link and panel for containers (#203)

* Fix panel link to pgBackRest dashboard

The dashboard uid is lowercase. Include variables so that the cluster
stays selected.

Issue: [ch9205]

* Fix pgBackRest backup runtime panel

The correct metric name is ccp_backrest_last_info_backup_runtime_seconds.

Issue: [ch9206]

Co-authored-by: tjmoore4 <42497036+tjmoore4@users.noreply.github.com>
Co-authored-by: Joseph Mckulka <16840147+jmckulk@users.noreply.github.com>
Co-authored-by: TJ Moore <tj.moore@crunchydata.com>
Co-authored-by: Keith Fiske <keith@keithf4.com>
Co-authored-by: Jonathan S. Katz <jkatz@users.noreply.github.com>
Co-authored-by: Chris Bandy <chris.bandy@crunchydata.com>