Add OOMKiller Alert
Linux has been tracking OOMKiller stats in /proc/vmstat since Kernel 2.6.36.
node_exporter also provides this statistic as part of standard exposed
metrics.

If a host hits an OOM state, it's likely that an alert should be
generated, so that it could be reviewed for hardware or system resource
faults.